after previous changes failed they were all reverted and slowly readded the code currently works and displays accuratly multiple notes however they are not displayed as the right length eveytthing is a crochet

+++ globals.cpp
@@ -12,7 +12,7 @@ bool getBMPReady = false;
 bool bpmReady = false;
 std::vector<float> sharedBuffer(BUFFER_SIZE, 0.0f);
 std::vector<std::pair<int,double>> sharedRealTimeList;
-std::vector<std::pair<int,int>> BPMTimeList;
+std::vector<std::pair<int,double>> BPMTimeList;
 
 const std::chrono::steady_clock::time_point START = std::chrono::steady_clock::now();
 using TimeDuration = std::chrono::duration<int64_t, std::nano>;
